Experiments_docker/cloud/requirements_cloud.txt
paho-mqtt
 
matplotlib==3.0 
numpy==1.16.4  
torch==1.1.0 
torchvision==0.3.0 
pillow==6.0.0
mpmath
